requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
selenium==4.4.3
